+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 - Fixes bug where GCS metadataUpdate cloud functions were triggered in incorrect situations (#3398).
 - Fixes bug where quoted escape sequences in .env files were incompletely unescaped. (#4270)
 - Fixes Storage Emulator ruleset file watcher (#4337).
+- Fixes issue with importing Storage Emulator data exported prior to v10.3.0 (#4358).